refactor: reorganize map features
