+++ common/src/main/java/com/wynntils/services/mapdata/MapFeatureService.java
@@ -11,9 +11,9 @@ import com.wynntils.core.components.Services;
 import com.wynntils.core.net.Download;
 import com.wynntils.core.net.UrlId;
 import com.wynntils.services.map.PoiService;
-import com.wynntils.services.mapdata.providers.builtin.CombatListProvider;
-import com.wynntils.services.mapdata.providers.builtin.PlaceListProvider;
-import com.wynntils.services.mapdata.providers.builtin.ServiceListProvider;
+import com.wynntils.services.mapdata.features.CombatLocation;
+import com.wynntils.services.mapdata.features.PlaceLocation;
+import com.wynntils.services.mapdata.features.ServiceLocation;
 import java.util.List;
 
 public class MapFeatureService extends Service {
@@ -28,8 +28,8 @@ public class MapFeatureService extends Service {
     private void loadPlaces() {
         Download dl = Managers.Net.download(UrlId.DATA_STATIC_MAPDATA_PLACES);
         dl.handleReader(reader -> {
-            TypeToken<List<PlaceListProvider.PlaceLocation>> type = new TypeToken<>() {};
-            List<PlaceListProvider.PlaceLocation> places = PoiService.GSON.fromJson(reader, type.getType());
+            TypeToken<List<PlaceLocation>> type = new TypeToken<>() {};
+            List<PlaceLocation> places = PoiService.GSON.fromJson(reader, type.getType());
             Services.MapData.PLACE_LIST_PROVIDER.updatePlaces(places);
         });
     }
@@ -37,8 +37,8 @@ public class MapFeatureService extends Service {
     private void loadServices() {
         Download dl = Managers.Net.download(UrlId.DATA_STATIC_MAPDATA_SERVICES);
         dl.handleReader(reader -> {
-            TypeToken<List<ServiceListProvider.ServiceLocation>> type = new TypeToken<>() {};
-            List<ServiceListProvider.ServiceLocation> services = PoiService.GSON.fromJson(reader, type.getType());
+            TypeToken<List<ServiceLocation>> type = new TypeToken<>() {};
+            List<ServiceLocation> services = PoiService.GSON.fromJson(reader, type.getType());
             Services.MapData.SERVICE_LIST_PROVIDER.updateServices(services);
         });
     }
@@ -46,8 +46,8 @@ public class MapFeatureService extends Service {
     private void loadCombatLocations() {
         Download dl = Managers.Net.download(UrlId.DATA_STATIC_MAPDATA_COMBAT_LOCATIONS);
         dl.handleReader(reader -> {
-            TypeToken<List<CombatListProvider.CombatLocation>> type = new TypeToken<>() {};
-            List<CombatListProvider.CombatLocation> combatLocations = PoiService.GSON.fromJson(reader, type.getType());
+            TypeToken<List<CombatLocation>> type = new TypeToken<>() {};
+            List<CombatLocation> combatLocations = PoiService.GSON.fromJson(reader, type.getType());
             Services.MapData.COMBAT_LIST_PROVIDER.updateCombats(combatLocations);
         });
     }
