+++ common/src/main/java/com/wynntils/services/mapdata/providers/builtin/LootChestsProvider.java
@@ -4,10 +4,8 @@
  */
 package com.wynntils.services.mapdata.providers.builtin;
 
-import com.wynntils.models.containers.type.LootChestTier;
-import com.wynntils.services.mapdata.providers.json.JsonMapLocation;
+import com.wynntils.services.mapdata.features.FoundChestLocation;
 import com.wynntils.services.mapdata.type.MapFeature;
-import com.wynntils.utils.mc.type.Location;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
@@ -33,15 +31,5 @@ public class LootChestsProvider extends BuiltInProvider {
 
     public static void registerFeature(FoundChestLocation foundChest) {
         PROVIDED_FEATURES.add(foundChest);
-    }
-
-    public static final class FoundChestLocation extends JsonMapLocation {
-        public FoundChestLocation(Location location, LootChestTier tier) {
-            super(
-                    "found-chest" + "-" + location.hashCode(),
-                    "wynntils:personal:found-chest:tier-" + tier.getWaypointTier(),
-                    null,
-                    location);
-        }
     }
 }
