+++ common/src/main/java/com/wynntils/services/mapdata/providers/builtin/ServiceListProvider.java
@@ -4,10 +4,8 @@
  */
 package com.wynntils.services.mapdata.providers.builtin;
 
-import com.wynntils.services.mapdata.providers.json.JsonMapAttributes;
-import com.wynntils.services.mapdata.providers.json.JsonMapLocation;
+import com.wynntils.services.mapdata.features.ServiceLocation;
 import com.wynntils.services.mapdata.type.MapFeature;
-import com.wynntils.utils.mc.type.Location;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
@@ -33,11 +31,5 @@ public class ServiceListProvider extends BuiltInProvider {
 
     private static void registerFeature(ServiceLocation location) {
         PROVIDED_FEATURES.add(location);
-    }
-
-    public static final class ServiceLocation extends JsonMapLocation {
-        public ServiceLocation(String featureId, String categoryId, JsonMapAttributes attributes, Location location) {
-            super(featureId, categoryId, attributes, location);
-        }
     }
 }
