+++ common/src/main/java/com/wynntils/services/mapdata/providers/builtin/WaypointsProvider.java
@@ -4,13 +4,10 @@
  */
 package com.wynntils.services.mapdata.providers.builtin;
 
-import com.wynntils.services.mapdata.providers.json.JsonMapAttributes;
-import com.wynntils.services.mapdata.providers.json.JsonMapLocation;
+import com.wynntils.services.mapdata.features.WaypointLocation;
 import com.wynntils.services.mapdata.type.MapFeature;
-import com.wynntils.utils.mc.type.Location;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Locale;
 import java.util.stream.Stream;
 
 public class WaypointsProvider extends BuiltInProvider {
@@ -34,16 +31,5 @@ public class WaypointsProvider extends BuiltInProvider {
 
     public static void registerFeature(WaypointLocation waypoint) {
         PROVIDED_FEATURES.add(waypoint);
-    }
-
-    public static final class WaypointLocation extends JsonMapLocation {
-        public WaypointLocation(Location location, String label, String subcategory, JsonMapAttributes attributes) {
-            super(
-                    "waypoint" + "-" + label.toLowerCase(Locale.ROOT).replaceAll("\\s", "_") + "-"
-                            + location.hashCode(),
-                    "wynntils:personal:waypoint" + (subcategory.isEmpty() ? "" : ":" + subcategory),
-                    attributes,
-                    location);
-        }
     }
 }
